restructuring and basic cluster support
